LayoutTests/proximity/window-property-expected.txt

Tests that the window.DeviceProximityEvent and window.onwebkitdeviceproximity properties are present.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS typeof window.DeviceProximityEvent == 'object' is true
PASS typeof window.DeviceProximityEvent == 'function' is false
PASS hasDeviceProximityEventProperty() is true
PASS 'DeviceProximityEvent' in window is true
PASS window.hasOwnProperty('DeviceProximityEvent') is true
PASS typeof window.onwebkitdeviceproximity == 'object' is true
PASS hasOnDeviceProximityProperty() is true
PASS 'onwebkitdeviceproximity' in window is true
PASS window.hasOwnProperty('onwebkitdeviceproximity') is true
PASS successfullyParsed is true

TEST COMPLETE
